% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_values.R
\name{get_timeseries_values}
\alias{get_timeseries_values}
\title{Prompt user for all needed inputs and then get values for the time series id.}
\usage{
get_timeseries_values()
}
\value{
A tibble with the columns returned by the script ki_timeseries_values.
}
\description{
Asks the user for details to narrow down on the timeseries and then returns time series values for given time series id and date range.
}
\examples{
\dontrun{
get_timeseries_values()
}

}
\author{
Daniel Post, MVCA, \email{dpost@mvc.on.ca}
}
