% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_timeseries_values.R
\name{ki_timeseries_values}
\alias{ki_timeseries_values}
\title{Get values for time series id or list of time series ids.}
\usage{
ki_timeseries_values(ts_id, start_date, end_date)
}
\arguments{
\item{ts_id}{Either: a single time series id or a vector of time series ids.
Time series ids can be found using the `ki_timeseries_list` function.}

\item{start_date}{A date string formatted "YYYY-MM-DD". Defaults to yesterday.}

\item{end_date}{A date string formatted "YYYY-MM-DD". Defaults to today.}
}
\value{
A tibble with following columns by default: Timestamp, Value, Units, station_name, stationparameter_name, ts_name, ts_id
}
\description{
Returns time series values for given time series id and date range.
}
\examples{
\dontrun{
ki_timeseries_values(
  ts_id = "1395042",
  start_date = "2015-12-01",
  end_date = "2018-01-01"
)
}

}
\author{
Ryan Whaley, \email{rdgwhaley@gmail.com}
}
