% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcaWISKI-package.R
\docType{package}
\name{mvcaWISKI-package}
\alias{mvcaWISKI}
\alias{mvcaWISKI-package}
\title{mvcaWISKI: A wrapper for querying the MVCA's WISKI Databases via the KiWIS API}
\description{
mvcaWISKI provides a simplified method for retrieving tidy data from the MVCA WISKI databases via KiWIS API.
}
\details{
A suggested workflow for using this package:
\itemize{
 \item Get time series data using \code{ki_timeseries_values()}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mvc.on.ca}
}

}
\author{
\strong{Maintainer}: Daniel Post \email{dpost@mvc.on.ca}

Authors:
\itemize{
  \item Ryan Whaley \email{rdgwhaley@gmail.com}
}

Other contributors:
\itemize{
  \item Sam Albers \email{sam.albers@gov.bc.ca} [contributor]
}

}
